(svn r10258) -Codechange: as we are now using int64 all over the place, it's better to use int64 variables in the string generating too instead of packing them into two int32s. -Fix: some displays of money were wrong.

+++ src/openttd.h
@@ -353,7 +353,7 @@ struct TileDesc {
 	StringID str;
 	Owner owner;
 	Date build_date;
-	uint32 dparam[2];
+	uint64 dparam[2];
 };
 
 struct ViewportSign {
